[circt-synth] Lower comb.add to Parallel-Prefix Adder  (#8457)
## Summary
An addition to the CombToAIG pass, reducing adders of width greater than 8 to a parallel-prefix adder. Currently implemented two flavours of parallel-prefix reduction tree.

## Architecture
- Width < 8 -> Ripple Carry Adder - minimal area
- 8 <=Width < 32 -> Parallel-Prefix Adder using Kogge-Stone tree - better delay, worse wiring congestion and area than Brent-Kung
- 32 <= Width -> Parallel-Prefix Adder using Brent-Kung tree - better area and wiring congestion, one additional logic level than Kogge-Stone

## Testing
Verified using circt-lec the correctness of the implementation for adders of width=1,2,...,64.

Included separate integration tests for adders of width 16 and 33. Expanding the test-suite here can quickly lead to timeouts given the increasing complexity of the verification challenge. Worth a discussion on future testing plans as more complex component architectures are added.