[ESI][Runtime] Fix FIFO connections through cosim
When cosimulation endpoints are instantiated, they see the ValidReady
converted signaling. This doesn't match what's in the manifest since the
connection request type includes the FIFO signaling. So the gRPC string
description of the endpoint doesn't match with the type ID from the
manifest. "Fix" this by removing the sanity check.