[ci] Change static build compiler to stable loc
The use of `clang` and `clang++` is not stable in Alpine if you are
installing a version of clang which is not the latest.  E.g., the recent
Alpine update to clang20 broke the static build because there is no longer
a `clang` or `clang++` binary installed.  Fix this, by switching to a
stable name that won't change, i.e., `clang-19` and `clang++-19`.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>