[LLHD] Fix deseq poisoning comb ops that don't depend on triggers (#8513)
Fix an issue in the Deseq pass where operations like
`comb.icmp ne %false, %false` get mistakenly marked as poison. This was
due to a missing check whether the operands are constant true or false.
This cause any comb op not supported in Deseq's truth table mapping to
be mistakenly poisoned if the op contained any `i1` operands that were
constant true or false.

Fixes #8512