[LLHD] Disable check if reset value is constant in Deseq
Remove the call to `isValidResetValue` in the Deseq pass when matching
the clocking and reset behavior described by a process against the
available register ops.

Resets are problematic since Verilog forces designers to describe them
as edge-sensitive triggers. This does _not_ match the async resets found
on almost all standard cell flip-flops, which are level-sensitive.
Therefore a pass lowering from Verilog-style processes to a structural
register such as `seq.compreg` would have to verify that the reset value
is a constant in order to make the mapping from edge-sensitive Verilog
description to level-sensitive standard cell valid.

In practice, designers commonly implement registers encapsulated in a
Verilog module, with the reset value being provided as a module input
port. This makes determining whether the input is a constant much more
difficult. Most commercial tools relax this constraint and simply map
the edge-sensitive reset to a level-sensitive one. This does not
preserve the semantics of the input, which is bad. Most synthesis tools
will then go ahead and fail during synthesis if a register's reset value
does not end up being a constant value.

This commit removes any checks whether a reset value is a constant, in
order to allow the circt-verilog pipeline to work with real designs. We
may want to reintroduce these checks in the future, possibly in some
different form.