[FIRRTL] Fix public copying in inject-dut-hier
Change the behavior of how publicness is handled in the inject-dut-hier
pass to never create _new_ public modules.  This has ramifications as
creating new public modules will create additional reference files as
requires by the FIRRTL ABI.  Instead, either keep the existing publicness
or move the existing publicness (when using hte `moveDut=true` flow.

Note: the `moveDut` flow violates the FIRRTL ABI as will create references
which are different than what the ABI says.  This is deemed acceptable
because the inject-dut-hier flow is supposed to go away and if a user is
opting into this flow, they need to recognize that they are doing weird,
non-standard things and they must adjust their build flow accordingly.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>