[FIRRTL] Move all annotations in inject-dut-hier
Change FIRRTL's inject-dut-hier pass to move all annotations onto the
wrapper [^1] when in `moveDut=true` mode.  Additionally, when in this
mode, the WRAPPER will be named using the `name` field of this pass'
controlling annotation.  Previously, this would only move the
`MarkDUTAnnotation` when operating in this mode.  This fixes an internal
issue where Object Model paths were not being updated when using this pass
in `moveDut=true` mode.

[^1]: This is using the terminology of the pass which views this as taking
all the logic inside the original DUT and shoving it into a wrapper
instantiated inside the DUT.  When in `moveDut=true` mode, it is more
natural to think of the pass as creating a wrapper around the original
DUT.  However, to adopt this alternative terminology would be modal.
I.e., "wrapper" would mean something different when in `movedDut=true`
mode which would be confusing.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>