[FIRRTL] Allow GC Companion multi-instantiation
Change the Grand Central View lowering pass to now allow for Grand Central
Companions to be multiply instantiated.  This should have technically been
possible for a long time [[1]] as long as there was a guarantee that there
was never going to be a leaf that was _outside_ the companion.

This is a pretty simple change to just loop over all instances and
extract (add a `lowerToBind` attribute) for the ones which are in the
design.  The other, subtle blocker is that without the `InstanceInfo`,
doing the check of "in the design" was tedious.  Now this is pretty
straightforward.

[1]: https://github.com/llvm/circt/commit/8b1ad061e0777704cfd4bb1c80196cb236307588

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>