[CI] Change cache cleanup threshold from 1/2 to 1/3
Sometimes cache sizes for Windows CI builds on main branch are smaller
compared to caches from other branches for unknown reasons. In that case,
a cache from a successful build could be accidentally removed when using
the 1/2 threshold. This commit reduces the threshold to 1/3 to be more
conservative and preserve successful build caches.