[FIRRTL] Fix Grand Central companion not-in-design (#8625)
Fix a regression introduced in e08f5d528 where a Grand Central companion
instantiated only under the test harness would cause the companion to be
instantiated in the Grand Central directory and not under the test harness
directory.  An early exit if no instances are in the design was
incorrectly dropped.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>