[ci] Fix issues in cleanup cache
Fix an issue where the appropriate permissions were not set in the
`uploadReleaseArtifacts.yml` job that are necessary, transitively for the
`cleanupCache.yml` job.

Remove a deleted `dry_run` parameter passed to the `cleanupCache.yml` job
from the UBTI-static job.

Don't use `bash` unnecessarily as it isn't installed by default in alpine.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>