[FIRRTL] Delete BlackBoxResourceFileNameAnno
Remove `BlackBoxResourceFileNameAnno` since this has not been load beagin
after it was removed from Chisel [[1]].  This annotation was used to
control the name of a file which was produced that listed all blackboxes
in the design.  Instead, use the default name always.

Note: I plan to remove this file entirely in a follow-on as this appears
to be completely unused.

[1]: https://github.com/chipsalliance/chisel/commit/70f5c7066b1e65cbb9d580703449acb5912a2b05

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>