Convert to CF before running LLHD inline. (#8667)
The LLHD inliner doesn't check if the parent op only supports single
block regions when inlining a call (while the function body of the call
may already be multiblock). This flips the lowering to CF to before LLHD
inline to avoid a common variant of this (inlining into SCF ops).