[FIRRTL] Make Lint Circuit Pass w/ parallelism
Convert the FIRRTL linting pass to be a `CircuitOp` pass with internally
handled parallelism.  This is done in preparation to add XMR linting which
requires an `InstanceInfo` analysis which needs to be on the `CircuitOp`.
There are two ways of doing this:

  1. Make the pass operate on the `CircuitOp` as this patch does.

  2. Use a cached parent analysis which requires a dedicated pass to
     compute the `InstanceInfo` analysis and then careful pass pipeline
     structuring.

(2): is used in one place in IREE and nowhere else.  Additionally, (1) was
suggested by multiple people in "Hallway Usabilty Testing"-type
discussions.  (Many people thought (2) was odd.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>