[FIRRTL] Refactor lint to use ODS constructor
Refactor FIRRTL's Lint pass to use the ODS-generated constructor as
opposed to hand-rolling one.  This avoids weirdness where the command-line
options have different defaults from what they are specified as in ODS and
avoids the need for a "Config" struct inside the linter's namespace (as
the ODS-generated one can be used).

h/t @uenoku for the suggestion.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>