[firtool] Default -lint-xmrs-in-design to false
Change the default value for `-lint-xmrs-in-design` to `false`.  This
provides at least one version where this linting is available, but
opt-in (which is backwards compatible) as opposed to opt-out (which is
not).  Additionally, for flows that are using `-enable-layers` and
`-sv-extract-test-code`, defaulting this to `true` creates issues where
you _need_ to turn it off because linting happens after layers are
inlined, but before they are extracted.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>