[Datapath] Add Datapath to SMT conversion pass (#8682)
To verify correctness of transformations involving datapath operations,
include a lowering from datapath to SMT. Will later be integrated into
circt-lec to enable verification of these operators. Each operator
satisfied a contract, rather than providing a precise semantics for
every operator. This is because the datapath operators return values in
redundant number representations, meaning there are many valid
implementations.

For example:
```mlir
%0:2 = datapath.compress %a, %b, %c : i8 [3 -> 2]
```
Will be verified by introducing free variables for each return value 
`(%0#0, %0#1)` then asserting that the sum of the associated free 
variables is equal to the sum of the inputs:
`assert(%0#0 + %0#1 == %a + %b + %c)`.

Whilst this is encoded as an assert it really represents an assumption
that must be satisfied by a valid implementation of datapath.compress.