[MooreToCore] Mark Seq dialect as legal
Add the Seq dialect as a legal dialect for the Moore-to-Core conversion.
This will allow the conversion to properly pass through `seq.to_clock`
in case the conversion is called on an IR that is already lowered to the
core dialects. This allows circt-verilog to accept core-level IR as
input.