[lsp] Depend on new MLIRRegisterAllDialects library.
Upstream MLIR recently carved out a new static library for this in:

https://github.com/llvm/llvm-project/commit/e68a20

This moved the `mlir::registerAllDialects` function to a new library,
which we need to depend on in circt-lsp-server. A similar change was
made upstream for cir-lsp-server, for example.

This appears to be our only use of `mlir::registerAllDialects`.