[FIRRTL][CreateSiFiveMetadata] Add memory read-under-write behavior to emitted metadata (#8604)
This PR adds information on the read under write behavior of memories, as specified by the FIRRTL/CHIRRTL IR, to the emitted memory metadata records. This includes the textual `repl-seq-mem-file`, the `seq_mem.json`, and the object model IR. It should allow external tools to generate memories with the appropriate behavior.

* In the JSON file, there is a new string attribute `ruw_behavior` that can take one of the following values : `Undefined`, `New`, `Old`.
* In the textual file, Undefined behavior is elided to keep the format unchanged for the common case. Otherwise `ruw New` or `ruw Old` is appended to the memory's line.
* In the object model it is added to the MemorySchema as a string-typed subfield called `ruwBehavior`.