[FIRRTL] Add domain information to types
Add domain information to some FIRRTL types: UInt, SInt, and Analog.  The
MLIR representation of FIRRTL Dialect now supports all of the following:

``` mlir
!firrtl.uint<1>
!firrtl.uint<1, []>
!firrtl.uint<1, [@A]>
!firrtl.uint<1, [@A, @B]>
```

Note: this should be extended in the future to be on all base types or on
all types.  When doing this extension, we will need to figure out how to
handle bundles where there are redundant encodings for types (which is a
huge red flag if allowed).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>