[FIRRTL] Canonicalize domains on scalar types
For scalar types which currently support domains (UInt, SInt, and Analog),
canonicalize their domains in their backing storage.  Specifically, sort
and unique the domains.  This means that domains will be stored
alphabetically and no duplicates will be allowed.

This is done to enable trivial pointer comparisons between types and not
require deep introspection to know if two types are the same.  This does
not, yet solve the problem of canonicalization involving aggregates whose
leaves may contain domains.

Aside: this will not run into the same problems that the mythological "flip
type canonicalization" effort ran into.  There are no operations which
care about the order of domains.  We are free to rearrange them as we want.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>