[Support] Use git commit hash as version in non-release builds (#8834)
Instead of just saying "unknown git version", use the current git commit
hash as CIRCT version in non-release builds. This makes the output of
`--version` of our tools more useful, because it points to an actual
commit.

With this change, the [sv-tests-results] website can provide an actual
git commit tooltip for the circt-verilog tool being used.

[sv-tests-results]: https://chipsalliance.github.io/sv-tests-results/