[FIRRTL] Add `unsafe_domain_cast` operation
Add a new operation, `firrtl.unsafe_domain_cast`, which is used to do
unsafe conversions to domains.  Conceptually, this acts like a node whose
input has unknown domain type and result has a known domain type.  This is
not a generic type ascription, as the domain to which the input is casted
does not affect the input at all.  The domain type is only assigned to the
result.

Add a sipmle canonicalizer that will replace an `unsafe_domain_cast` which
has no domains specified.  This operation is a no-op and there's no point
in keeping it around.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>