[HW] Accept most reductions even if module size increases (#8847)
The HW reductions that replace values with constants and remove module
ports are beneficial even when they increase the size of the MLIR text.
Removing ports and replacing values with zero removes dependencies in
the IR, making it easier to apply other reductions.

Make these reductions return true from `acceptSizeIncrease` to have the
reducer accept the result even if the module size increases.