[ImportVerilog] Error on range select with non-constant bounds
Produce a proper error when encountering a range select with a dynamic
right-hand side of the range. We only support constant ranges like
`[a:b]` or ranges with a dynamic offset but constant length like
`[a+:42]` at the moment. Instead of triggering an assertion if the user
types something like an unbounded range on a queue, e.g. `[2:$]`, emit a
proper error diagnostic.