[Comb] Avoid infinite loops by not canonicalizing recursive ops (#8866)
Add a `isOpTriviallyRecursive` helper function as originally proposed by
@fzi-hielscher and use it to avoid folding and canonicalizing ops that
are trivially recursive. This avoids creating infinite loops in the
canonicalizer when recursive structures are present in the input.

This is pretty conservative, since certain recursive structures like
`%0 = comb.and %0, %c0` *could* be folded. We may want to opt into
allowing these folds explicitly in the future. For now, it's better to
not have a crashing/hanging compiler.

Note that these inputs are very rare. Usually they come up due to
SystemVerilog combinational processes describing loopy dataflow, or the
circt-reduce tool trying to minimize some test case and forwarding op
operands in an unfortunate fashion.