[ArcToLLVM] Replace all i0 values with a constant 0 (#8871)
I have noticed that the emission to LLVMIR doesn't like `i0` types in
the IR, leading to assertions in `llvm::IntegerType` to fail. Add a
preprocessing step to the Arc-to-LLVM conversion pass that replaces all
`i0` values in the IR with an `hw.constant 0 : i0`. This causes the `i0`
values in the LLVM lowering to disappear, with most of them getting
constant-folded from `%1 = llvm.zext %0 : i0 to i42` to
`%1 = llvm.mlir.constant(0 : i42) : i42`.