[Moore] Add more precise conversion ops (#8883)
Extend the Moore dialect with the following new conversion operations:

- `moore.packed_to_sbv` to convert any packed type to its simple bit
  vector equivalent (i.e., and integer);
- `moore.sbv_to_packed` to do the reverse;
- `moore.logic_to_int` to convert from a four-valued integer to a
  two-valued integer;
- `moore.int_to_logic` to convert from a two-valued integer to a
  four-valued integer; and
- `moore.to_builtin_bool` to convert from a single bit `moore::IntType`
  to a single bit builtin `mlir::IntegerType`.

Adjust the `ImportVerilog` pass to generate these new conversion ops
instead of the old whole-sale `moore.conversion`. Having smaller, more
precise conversion ops will make lowering to the core dialects easier in
the future, especially once we fully support four-valued types and time
values.

In the future, we may also want to create conversion ops from and to
some of the unpacked types such as strings, unpacked arrays, queues,
etc. These will require special care in the lowering, likely expanding
to runtime function calls.

The long-term plan is to get rid of `moore.conversion` altogether and
replace it with simpler, dedicated conversion ops.

Most of this commit is just mechanically threading-through the new cast
operations. They all lower to the same bitcast (or a passthrough) in HW
as the `moore.conversion` op. This should result in no effective change
at the core dialect level.

I ran this on [circt-tests] and the Snitch core in [arc-tests], with no
change in output.

[circt-tests]: https://github.com/circt/circt-tests
[arc-tests]: https://github.com/circt/arc-tests