[Sim] Create clocked and procedural sim.exit/pause ops (#8893)
Combine the existing `sim.finish` and `sim.fatal` ops into a single
`sim.clocked_exit` op. Create a `sim.exit` op for the procedural case.
This op now accepts a `success` operand to indicate whether the
simulation should terminate with a success or failure exit code, and a
`verbose` operand to indicate whether the exit should behave like a
`$finish(0)` or a `$finish(1)` in SystemVerilog.

Also add a `sim.pause` and `sim.clocked_pause` operation which only
suspends simulation and gives control back to the user in an interactive
simulation. This corresponds to `$stop` in SystemVerilog.

These new ops will allow circt-verilog to lower `$finish`, `$fatal`,
`$stop`, and `$exit` in the input.

Adjust the Sim-to-SV conversion pass to support the new operations. Also
try to reuse existing ifdef and always/if ops, effectively combining
multiple consecutive Sim ops into a single ifdef guard or always
process.