[Synth] Refactor and improve cut rewriter infrastructure (#8909)
This commit refactors the cut rewriter and implements correct priority sort metric for depth. Previously `depth` of a cut represents a logic depth *within* a cut but this is not correct. What the original paper (ICCAD 2007) proposed was to use arrival time of cut outputs instead.

For that this commit changes to perform cut matching for each cut, and use worst arrival time as a priority of cuts.

LUT6 depth for 16-bit multiplier improved from 17 to 7, which is on per with Yosys/ABC.