[ESI] Don't swallow exception in callback port wrapper
I'm open to convincing, but I don't think the type conversion wrapper should swallow an exception occuring during argument deserialization and callback handling. Users of the ESI runtime may have other logging mechanisms active, so catching and printing an exception through `traceback` doesn't make sense to me.

This PR removes that, and assumes that the user of the callback port is able to handle exceptional cases.