[ESI] Factor out inner execution in Simulator::run (#8964)
* [ESI] Factor out inner execution in Simulator::run

When using esi-cosim as a library, it is very inconvenient that the default behavior of the `run` command assumes that `esi-cosim` is executed as a standalone script (i.e. either tries to run a user-provided command or waits for user input).
To better serve as a library, factor out the inner simulator process execution to `run_proc` s.t. other python code can call this, and get a handle to the simulator process and port.

* factor out finally block