[ConvertToArcs] Add llhd.combinational conversion (#8950)
Introduce a dialect conversion step into the `ConvertToArcs` pass. We'll
use this to map various core dialect operations to Arc-specific ones in
the future. As a starting point, add a conversion from
`llhd.combinational` to `arc.execute`. This, alongside lowering to LLVM
to be added in later PRs, will implement some basic LLHD support in the
Arc dialect.