[Moore] Fix missing RValue conversion for struct_create (#8988)
This PR fixes a bug where RValue Conversion is not called when a struct
is created from immediate values.

Per IEEE 1800-2023 Section 10.9.2 "Structure assignment patterns", the
following assignment pattern is valid:

```
structType st = '{member: value, memberTwo: valueTwo, ...};
```

In ImportVerilog, this pattern is recognized and mapped to a 
`StructCreateOp` operator.

Until now the RValue expressions for value and valueTwo were converted
without a target type. This leads to a failure to convert e.g. between
ArrayType instances and IntType instances. This PR fixes this by
passing the structType member types to the convertRValueExpression
function to have them be cast explicitly.