[Synth]LowerWordsToBits: Improve scalability with bit-sensitive constprop (#8997)
Previously, GreedyPatternRewriter was naively used to bit-blast multi-bit values but it didn't scale well due to worklist iteration and unnecessary IR creation which occasionally ended up with constant values. Hence the pass was often not scalable for large designs.

This commit fixes the issue by applying constant propagation during the bit-blast process and replaces GreedyPatternRewriter with a manual one-shot walk with recursion.