[Arc] Add MergeTaps pass (#9000)
The MergeTaps pass collects `arc.tap` operations that observe the same (SSA) value and merges them into a single TapOp if they are within the same HWModule body and not nested under another region. To do so, the TapOp itself had to be changed to carry a StrArrayAttr for the names instead of just a single string.