[FIRRTL] Remove class deduplication
The FIRRTL Dedup pass currently also deduplicates classes. This is
problematic, since classes are nominally typed, which requires their
names to be fixed and doesn't allow us to swap one class for another
even if they are structurally equivalent. Downstream tools expect the
classes to retain their names, which is problematic if they look for
class "Foo", but "Foo" happened to be structurally equivalent to "Bar"
and therefore got replaced with "Bar".

To fix this, remove any class handling from the dedup pass. It can no
longer remove classes or change an object's class. As a result, we can
now treat class types as opaque again, since the class name symbols they
contain will not be changed by the pass.

An alternative approach would have been to remove the nominal typing
property from classes. If names are not stable, the compiler would be
free to arbitrarily deduplicate and rework classes. This would imply
changes in downstream tools, and @mikeurbach pointed out that the intent
behind FIRRTL's classes were for them to be nominally typed. This now
puts the onus of not generating gratuitous class definitions on Chisel
and other frontends that produce FIRRTL.