[Support] Add RAII Debug Header/Footer helper
Add an resource-allocation-is-initialization style utility that will print
a debug header on construction and footer on destruction.  This better
encapsulates a common pattern in some passes (mostly FIRRTL passes) of
doing this manually.

Note: because this is  RAII, this needs to use ifdefs and not LLVM_DEBUG
to implement it.

h/t @dtzSiFive for the suggestion.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>