[Support] Improve scoped pass logger, add macro
Improve/fix the ScopedDebugPassLogger so that this can be used with RAII
patterns (as it was always intended to be!).  Use upstream LLVM debug
functions that allow for querying the debug state.  Don't, however, use
DEBUG_TYPE as this is tedious to properly thread through.  Instead,
directly query the pass' name and use this to set debugging.  Note: this
requires that the specified DEBUG_TYPE correctly matches the pass name and
is not something different.

In the future, add macros to help with properly setting the DEBUG_TYPE for
a pass.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>