[FIRRTL] Add fields to domains
Extend domains to allow for them to contain fields.  These fields provide
a schema of information that all domains of certain kind must have.  This
models information that a user must provide for an input domain or is
provided for an output domain.

Opt for a "struct-like" representation as opposed to using input/output
ports like FIRRTL classes.  That said, these are intended to be lowered to
classes in a future patch.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>