[Synth] Add MaximumAndCover Pass (#9090)
This implementation is based on a tree-balancing algorithm described in section 3.1 in the paper: "Delay Optimization Using SOP Balancing, ICCAD 2011".

This pass performs maximum AND-cover optimization by collapsing single-fanout and-inverter ops into their users, reducing the depth of And-Inverter Graphs, combined with the lower variadic pass. The pass is integrated into the circt-synth optimization pipeline. 

Rebased from #8262.
Co-authored-by: Max Zhou <651535280@qq.com>