[OM] Add UnknownValue (in evaluator only)
Add the ability to represent an "unknown" value.  This is intended to be
used by the evaluator to deal with placeholders whose value the evaluator
user doesn't care about.  These unknown values are necessary in order to
allow for the "partial" evaluation of an object allowing a user to set
some object inputs to "unknown" and only set the ones they care about.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>