[OM] Support UnknownValue in Evaluator
Add support to the Evaluator to handle the new "unknown" values.  These
are values that the Evaluator user doesn't care about.  Any interaction
with these values results in the propagation of an "unknown" value.  (They
are poison-like.)  This enables evaluator users to _not_ have to specify
all values to a class they want to instantiate.  Instead, they can specify
only the inputs they care about and then read out any outputs that are
_not_ "unknown".

More conceptually, there is a problem with the design of OM that this is
working around.  We'd really like to think about an OM class as having
some number of functions that a user can provide parameters to and get
outputs.  However, the current design only allows for exactly one
function.  If we don't have the ability to do "partial" evaluation of this
function, then we need to provide values for parameters that we really
don't care about.

Note: "partial" is intentionally in quotes as this is not true partial
evaluation (in a functional programming) sense.  Just superficially
related.  The result of "partial" evaluation is not another function, it's
a "partial" result.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>