[FIRRTL] Parse domainless MLIR into empty array
Fix a subtle discrepancy between the FIRRTL text parser and the custom
FIRRTL MLIR parser where modules without domain information in the former
would be stored as an empty array and in the latter would be stored as an
array of empty arrays.  Correct the MLIR parser to use the more optimal
empty array representation.

Add tests that this behavior works for both domains and port
annotations.  (Port annotations were the first to use this
representation.)

Note: this decision for redundant storage representations should be
revisited in the future as it makes it tedious for downstream passes to
have to handle both representations.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>