Revert "[ci] Change static build to use Rocky Linux 8"
This reverts commit 081ff8f5bd24113d7ce2c086ef646ebabe493223.

Thanks to the fix by @fzi-hielscher in a1ca42556, the muslc builds will
continue to work, by not running Arcilator JIT tests.  We think that there
is line of sight to enabling these, though.

In the future, we may revisit switching to Rocky Linux for publishing.
However, we don't need to do this right now.