[LTL] Add Implication Folder
Fold implication of a constant false antecedent or constant true
consequent to true.  This is a pattern which has been empirically observed
in Chisel-generated FIRRTL code and this is moving towards fixing it.

This improves #9277, but does not go quite far enough.  This will get that
to:

```
assert property (disable iff (_GEN) 1'h1);
```

This can be later cleaned up to remove trivially true assertions (which
this is) now that we have the Boolean Constant Op.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>