[Verif] Lower simulation tests to HW modules (#9294)
Rename the *LowerFormalToHW* pass to *LowerTests* and make it handle
`verif.simulation` ops in the same way it handles `verif.formal`. The
only difference being the ports needed on the module, and the lack of
symbolic value ops.

Also add a comment to the modules generate by this class to make finding
them after Verilog emission easier. With these changes, a user can grep
for "// SIMULATION TEST" or "// FORMAL TEST" across the generated
Verilog files to find lit-style run lines for each test.