[Verif] Add more AssertLike canonicalizers (#9281)
Add some more, missing canonicalizers of asserts, assumes, and covers.
Specifically:

1. If an assert/assume/cover is always enabled, remove the optional
   enable.
2. If an assert/assume is always disabled, delete it.
3. If an assert/cover's condition is an always true (new)
   BooleanConstantOp, then delete it.

This technically fixes #9277.  There is still a missing improvement where
binds of empty modules can be dropped entirely because, it is impossible
for them to have any effect (or to be the target of other binds).

Note: the decision to not optimize covers aggressively is intentional.
Covers are part of the "verification contract" or "test plan" of a module.
These often are tracked from design inception through sign-off.  If we
remove trivially true covers, this destroys this traceability when we'd
often like to know that something was covered, even if
trivially.  (Additionally, removing these will change the total number of
covers which can cause coverage metrics to bounce around.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>