[MooreToCore] Added real_to_int support (#9346)
Added support for `real_to_int`. Since the standard doesn't concretely
specify the semantics of said conversion, for now, we choose to convert
the `RealType` value to a signed integer, with the same bitwidth as the 
lvalue. Also added corresponding test. Also added `FloatType` to the 
list of type conversions. In its absence, functions with arguments of 
type `moore::RealType` (`!moore.f32` and `!moore.f64`) couldn't legalize 
properly.