[ESI] Add type ser/de capability
This PR adds type ser/de capabilities for ESI runtime types. The ser/de format is inspired/copied from what is already now used in the manifest generation, although we do not use CIRCT type mnemonics, nor do we include CIRCT dialects in the serialized format. The intention is thus to _only_ map ESI runtime types.

If we indeed want a method for users to provide third-party types, the format used here could easily be extended to allow for external ser/de plugins to be hooked into these two functions.

All types are registered on the context upon construction. Likewise, we use the context as a cache for pre-existing types.