[Seq] Update HWMemSimImpl's replSeqMem to support higher latency.
Historically, we have only replaced FIRRTL memories with readLatency
and writeLatency of 1 with external modules for macro
replacement. The latency is an intrinsic behavior of the memory, but
doesn't affect its interface when we are doing external module
replacement.

Since some users are trying to enable higher latency memory models,
and replace them with external modules that either model this in
simulation or with a macro for synthesis. This lifts the restriction,
allowing such higher latency memories to also be replaced by external
modules similar to memories with latency of 1.