[FIRRTL] Fix LowerTypes so it works with Domains
Fix a bug in `LowerTypes` where domains were not properly being updated.
This is slightly tedious as `LowerTypes` does not use
`insertPorts`/`erasePorts` member functions as it uses the ports
themselves as a worklist.  Hence, this needs to use a custom helper,
`DomainLoweringHelper`, to reduce duplication between the lowering of
modules, extmodules, and instances.

Note: this hints that there is a lot of duplication between the lowering
of modules, extmodules, and instances that should be cleaned up.  However,
this is out of scope of this patch.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>