[FIRRTL] Quote IntegerAttr domain error, NFC
Quote the `IntegerAttr`s reported when an instance op doesn't have the
same domain info as its reference module op.  This is just slightly better
to read.

Note: it would be even nicer to just report the number.  However, these
need to match exactly and it is useful to directly report the type, e.g.,
signed vs. unsigned.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>