[FIRRTL] Relax DomainDefineOp wire verification
Weaken the verification involving `DomainDefineOp` to do no checking when
a `WireOp` is the source or destination.  This is not ideal, but is
necessary to allow for FIRRTL's module inlining pass to continue to
properly function.

When a module with domain ports is inlined, this creates wires for its
ports and drive them with `DomainDefineOp`.  This is problematic as the
created `WireOp`s do not support any domain information and this will then
trip the unmodified verifier here.

Note: the change in this patch can be rolled back with a proper fix to add
domain information to wires [[1]].  Additionally, an associated XFAIL test
has been added to track this.

[[1]]: https://github.com/llvm/circt/issues/9398

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>